+++ tests/blueant/test_connector.py
@@ -50,12 +50,14 @@ def test_get_json_from_api_mocked_error(mocked_blueant: MagicMock) -> None:
         connector._get_json_from_api("foo")
 
 
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_initialization() -> None:
     connector = BlueAntConnector.get()
     assert str(connector.session.headers["Authorization"]).startswith("Bearer")
 
 
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_get_projects() -> None:
     connector = BlueAntConnector.get()
@@ -93,6 +95,7 @@ def test_get_projects_mocked(mocked_blueant: BlueAntConnector) -> None:
     assert projects[0] == BlueAntProject.model_validate(project_dict["projects"][0])
 
 
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_get_client_name() -> None:
     connector = BlueAntConnector.get()
@@ -122,6 +125,7 @@ def test_get_client_name_mocked(mocked_blueant: BlueAntConnector) -> None:
     assert name == expected_name
 
 
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_get_type_description() -> None:
     connector = BlueAntConnector.get()
@@ -151,6 +155,7 @@ def test_get_type_description_mocked(mocked_blueant: BlueAntConnector) -> None:
     assert name == expected_description
 
 
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_get_status_name() -> None:
     connector = BlueAntConnector.get()
@@ -180,6 +185,7 @@ def test_get_status_name_mocked(mocked_blueant: BlueAntConnector) -> None:
     assert name == expected_description
 
 
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_get_department_name() -> None:
     connector = BlueAntConnector.get()
@@ -207,6 +213,7 @@ def test_get_department_name_mocked(mocked_blueant: BlueAntConnector) -> None:
     assert name == expected_name
 
 
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_get_persons() -> None:
     connector = BlueAntConnector.get()