+++ tests/confluence_vvt/test_transform.py
@@ -26,6 +26,7 @@ from mex.extractors.primary_source.helpers import (
 TEST_DATA_DIR = Path(__file__).parent / "test_data"
 
 
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_transform_confluence_vvt_page_to_extracted_activity(
     unit_merged_ids_by_synonym: dict[str, MergedOrganizationalUnitIdentifier],
@@ -73,6 +74,7 @@ def test_transform_confluence_vvt_page_to_extracted_activity(
     )
 
 
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_transform_confluence_vvt_page_to_extracted_activities(
     unit_merged_ids_by_synonym: dict[str, MergedOrganizationalUnitIdentifier],