+++ tests/datscha_web/test_connector.py
@@ -3,6 +3,7 @@ import pytest
 from mex.extractors.datscha_web.connector import DatschaWebConnector
 
 
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_datscha_connector_item_urls() -> None:
     datscha = DatschaWebConnector.get()
@@ -11,6 +12,7 @@ def test_datscha_connector_item_urls() -> None:
     assert len(list(item_urls)) > 10
 
 
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_datscha_connector_get_item() -> None:
     datscha = DatschaWebConnector.get()