+++ tests/drop/test_drop.py
@@ -8,6 +8,7 @@ from mex.extractors.drop import DropApiConnector
 
 
 @pytest.mark.skip("Mex-drop might not be running. To be fixed in MX-1576")
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_list_files() -> None:
     connector = DropApiConnector.get()
@@ -16,6 +17,7 @@ def test_list_files() -> None:
 
 
 @pytest.mark.skip("Mex-drop might not be running. To be fixed in MX-1576")
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_list_files_of_non_existing_x_system() -> None:
     connector = DropApiConnector.get()
@@ -37,6 +39,7 @@ def test_list_files_mocked(monkeypatch: MonkeyPatch) -> None:
 
 
 @pytest.mark.skip("Mex-drop might not be running. To be fixed in MX-1576")
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_get_file() -> None:
     connector = DropApiConnector.get()
@@ -45,6 +48,7 @@ def test_get_file() -> None:
 
 
 @pytest.mark.skip("Mex-drop might not be running. To be fixed in MX-1576")
+@pytest.mark.integration
 @pytest.mark.requires_rki_infrastructure
 def test_get_file_that_does_not_exist() -> None:
     connector = DropApiConnector.get()